Add basic rendering support
With some o!f changes and VALIDATE_USAGES disabled, it runs to clearing
my screen.

- Had to downgrade to 2.20.0 to fix a broken native call in
`wgpuCommandEncoderBeginRenderPass`. Something along the lines of
"invalid load op: 0". Doesn't look to be reported upstream so it's a
matter of finding the right moment to do so.

- Removed dependency on Dawn. Silk.NET for some reason released Dawn
bindings without actually releasing a dawn native lib. /shrug we don't
actually seem to require it.

- The removal of Dawn changes the way the swapchain textures are done a
bit.